feature(Pomo API):
Added:
- API Support for Pomos
- Pomo API tests

Modified:
- Pomo Policy (Still needs to be fixed)

+++ app/Http/Controllers/PomoController.php
@@ -4,7 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StorePomoRequest;
 use App\Http\Requests\UpdatePomoRequest;
+use App\Http\Resources\PomoResource;
 use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
+use Illuminate\Support\Facades\Auth;
 use App\Models\{
     Pomo,
     Project,
@@ -18,8 +23,15 @@ class PomoController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request): PomoResource|Application|Factory|View
     {
+        if ($request->expectsJson()){
+            $todo = Todo::find($request->todo_id);
+            $pomos = $todo->pomo()->paginate(4);
+
+            return new PomoResource($pomos);
+
+        }
         return view('pomo.index');
     }
 
@@ -42,7 +54,7 @@ class PomoController extends Controller
      */
     public function store(StorePomoRequest $request)
     {
-//        $this->authorize('create', Pomo::class);
+        $this->authorize('create', Pomo::class);
 
         // Convert due_start and end to unix timestamp and save
         $pomo = new Pomo();
@@ -52,6 +64,13 @@ class PomoController extends Controller
         $pomo->notes = $request->safe()->notes;
         $pomo->save();
 
+        if ($request->expectsJson()){
+            return response()->json([
+                'message' => 'Pomo created successfully.',
+                'data' => $pomo,
+            ], 201);
+        }
+
         return redirect()->route('pomo.index')
             ->with('success', 'Pomo created successfully.');
     }
@@ -59,9 +78,16 @@ class PomoController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Pomo $pomo)
+    public function show(Request $request, Pomo $pomo)
     {
         $this->authorize('view', $pomo);
+
+        if ($request->expectsJson()){
+            return response()->json([
+                'message' => 'Pomo retrieved successfully.',
+                'data' => $pomo,
+            ], 200);
+        }
 
         return view('pomo.show', compact('pomo'));
     }
@@ -84,13 +110,20 @@ class PomoController extends Controller
      */
     public function update(UpdatePomoRequest $request, Pomo $pomo)
     {
-//        $this->authorize('update', $pomo);
+        $this->authorize('update', $pomo);
 
         // Convert due_start and end to unix timestamp and save
         $pomo->pomo_start = strtotime($request->pomo_start);
         $pomo->pomo_end = strtotime($request->pomo_end);
         $pomo->notes = $request->notes;
         $pomo->save();
+
+        if ($request->expectsJson()){
+            return response()->json([
+                'message' => 'Pomo updated successfully.',
+                'data' => $pomo,
+            ], 200);
+        }
 
         return redirect()->route('pomo.index')
             ->with('success', 'Pomo updated successfully.');
@@ -100,12 +133,18 @@ class PomoController extends Controller
      * Remove the specified resource from storage.
      * @throws AuthorizationException
      */
-    public function destroy(Pomo $pomo)
+    public function destroy(Request $request, Pomo $pomo)
     {
         // Validate that the user is authorized to delete the pomo
-//        $this->authorize('delete', $pomo);
+        $this->authorize('delete', [Pomo::class, $pomo]);
 
         $pomo->delete();
+
+        if ($request->expectsJson()){
+            return response()->json([
+                'message' => 'Pomo deleted successfully.',
+            ], 200);
+        }
 
         return redirect()->route('pomo.index')
             ->with('success', 'Pomo deleted successfully.');
